WIP: layer gpu distribution working but val acc not right
